Refactored to fix race-condition when acknowledging events.
Main reason was, that we keept a pointer to amqp.Delivery in the amqpEvent. Which was updated in the
amqp_subscriber.go#runEventTransformer loop for each message from rabbitmq. Thus, if the transformer
ran twice, and then the amqp_consumer from the stats service ran twice, the same message would have
been acknowledged twice. Thus the error.

This commit also moves the consumerId and queueName() function into amqp_subscriber.go, because I
initially assumed we ran into problem when we resubscribe to the same channel with the same
consumerId. Thus we now recreate it for every Consume()